% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{all_fits_analysis}
\alias{all_fits_analysis}
\title{Run model parameter estimation analysis and plot results. This script analyses
all fits.}
\usage{
all_fits_analysis(model, df, plots_dir, data_point_num,
  fileout_param_estim_details, fileout_param_estim_summary,
  plot_2d_66cl_corr = FALSE, plot_2d_95cl_corr = FALSE,
  plot_2d_99cl_corr = FALSE, logspace = TRUE, scientific_notation = TRUE)
}
\arguments{
\item{model}{the model name without extension}

\item{df}{the dataset containing all the parameter estimation fits.}

\item{plots_dir}{the directory to save the generated plots}

\item{data_point_num}{the number of data points used for parameterise the model}

\item{fileout_param_estim_details}{the name of the file containing the detailed statistics for the estimated parameters}

\item{fileout_param_estim_summary}{the name of the file containing the summary for the parameter estimation}

\item{plot_2d_66cl_corr}{true if the 2D parameter correlation plots for 66\% confidence intervals should be plotted. This can be time consuming. (default: FALSE)}

\item{plot_2d_95cl_corr}{true if the 2D parameter correlation plots for 95\% confidence intervals should be plotted. This can be time consuming. (default: FALSE)}

\item{plot_2d_99cl_corr}{true if the 2D parameter correlation plots for 99\% confidence intervals should be plotted. This can be time consuming. (default: FALSE)}

\item{logspace}{true if parameters should be plotted in logspace. (default: TRUE)}

\item{scientific_notation}{true if the axis labels should be plotted in scientific notation (default: TRUE)}
}
\description{
Run model parameter estimation analysis and plot results. This script analyses
all fits.
}
