% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{all_fits_pe}
\alias{all_fits_pe}
\title{A parameter estimation data set including all the evaluated fits.}
\format{A data frame with 4 variables:
\describe{
\item{ObjectiveValue}{the objective value for this parameter set}
\item{k1}{The first estimated parameter}
\item{k2}{The second estimated parameter}
\item{k3}{The third estimated parameter}
}}
\usage{
all_fits_pe
}
\description{
Estimated parameters for a mini model of the insulin receptor.
}
\keyword{datasets}
