% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_ggplot2_themes.r
\name{basic_theme}
\alias{basic_theme}
\title{A generic basic theme for time courses. It extends ggplot2 theme_classic().}
\usage{
basic_theme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{the font size}

\item{base_family}{the font family}
}
\description{
A generic basic theme for time courses. It extends ggplot2 theme_classic().
}
\examples{
library(ggplot2)
theme_set(basic_theme(36))
}
