% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{best_fits_pe}
\alias{best_fits_pe}
\title{A parameter estimation data set including only the best evaluated fits.}
\format{A data frame with 4 variables:
\describe{
\item{Estimation}{the number of estimated parameter sets}
\item{ObjectiveValue}{the best objective value for this parameter estimation}
\item{k1}{The first estimated parameter}
\item{k2}{The second estimated parameter}
\item{k3}{The third estimated parameter}
}}
\usage{
best_fits_pe
}
\description{
Estimated parameters for a mini model of the insulin receptor.
}
\keyword{datasets}
