% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_sim.r
\name{check_exp_dataset}
\alias{check_exp_dataset}
\title{Check the experimental data set.}
\usage{
check_exp_dataset(exp_dataset, plot_exp_dataset = FALSE)
}
\arguments{
\item{exp_dataset}{a full path file containing the experimental data.}

\item{plot_exp_dataset}{TRUE if the data set file should be plotted.}
}
\value{
TRUE if the file exists.
}
\description{
Check the experimental data set.
}
