% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{combine_param_best_fits_stats}
\alias{combine_param_best_fits_stats}
\title{Combine the parameter best fits statistics.}
\usage{
combine_param_best_fits_stats(plots_dir, fileout_param_estim_best_fits_details)
}
\arguments{
\item{plots_dir}{the directory to save the generated plots}

\item{fileout_param_estim_best_fits_details}{the name of the file containing the statistics for the parameter best fits.}
}
\description{
Combine the parameter best fits statistics.
}
