% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{combine_param_ple_stats}
\alias{combine_param_ple_stats}
\title{Combine the statistics for the parameter estimation details}
\usage{
combine_param_ple_stats(plots_dir, fileout_param_estim_details)
}
\arguments{
\item{plots_dir}{the directory to save the generated plots}

\item{fileout_param_estim_details}{the name of the file containing the detailed statistics for the estimated parameters}
}
\description{
Combine the statistics for the parameter estimation details
}
