% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{compute_bic}
\alias{compute_bic}
\title{Compute the Bayesian Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2}
\usage{
compute_bic(chi2, k, n)
}
\arguments{
\item{chi2}{the Chi^2 for the model}

\item{k}{the number of model parameters}

\item{n}{the number of data points}
}
\value{
the BIC
}
\description{
Compute the Bayesian Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2
}
