% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{compute_cl_objval}
\alias{compute_cl_objval}
\title{Compute the confidence level based on the minimum objective value.}
\usage{
compute_cl_objval(min_objval, params, data_points, level = 0.5)
}
\arguments{
\item{min_objval}{the minimum objective value}

\item{params}{the number of parameters}

\item{data_points}{the number of data points}

\item{level}{the confidence level threshold (e.g. 0.1, 0.5)}
}
\value{
the confidence level based on minimum objective value
}
\description{
Compute the confidence level based on the minimum objective value.
}
