% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_sim.r
\name{compute_descriptive_statistics}
\alias{compute_descriptive_statistics}
\title{For each time point compute the most relevant descriptive statistics: mean, sd, var, skew, kurt, ci95, coeffvar,
min, 1st quantile, median, 3rd quantile, and max.}
\usage{
compute_descriptive_statistics(timepoint.values, nfiles)
}
\arguments{
\item{timepoint.values}{array of values for a certain time point}

\item{nfiles}{the number of files (samples)}
}
\value{
the statistics for the array of values for a specific time point
}
\description{
For each time point compute the most relevant descriptive statistics: mean, sd, var, skew, kurt, ci95, coeffvar,
min, 1st quantile, median, 3rd quantile, and max.
}
