% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{compute_fratio_threshold}
\alias{compute_fratio_threshold}
\title{Compute the fratio threshold for the confidence level.}
\usage{
compute_fratio_threshold(m, n, p = 0.05)
}
\arguments{
\item{m}{number of model parameters}

\item{n}{number of data points}

\item{p}{significance level}
}
\value{
the f-ratio threshold
}
\description{
Compute the fratio threshold for the confidence level.
}
\examples{
compute_fratio_threshold(5, 100)
compute_fratio_threshold(5, 100, p=0.01)
}
