% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{final_fits_analysis}
\alias{final_fits_analysis}
\title{Run model parameter estimation analysis and plot results. It analyses
only the best fits using a percent threshold.}
\usage{
final_fits_analysis(model, df, plots_dir, best_fits_percent, logspace = TRUE,
  scientific_notation = TRUE)
}
\arguments{
\item{model}{the model name without extension}

\item{df}{the dataset containing the best parameter estimation fits.}

\item{plots_dir}{the directory to save the generated plots}

\item{best_fits_percent}{the percent of best fits to analyse.}

\item{logspace}{true if parameters should be plotted in logspace.}

\item{scientific_notation}{true if the axis labels should be plotted in scientific notation (default: TRUE)}
}
\description{
Run model parameter estimation analysis and plot results. It analyses
only the best fits using a percent threshold.
}
