% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_sim.r
\name{gen_stats_table}
\alias{gen_stats_table}
\title{Plot model readouts with statistics for each time point.}
\usage{
gen_stats_table(inputdir, outputdir, model, outputfile, xaxis_label = "",
  yaxis_label = "", column_to_read = "X1")
}
\arguments{
\item{inputdir}{the input directory containing the time course files}

\item{outputdir}{the output directory}

\item{model}{the model name}

\item{outputfile}{the name of the file to store the statistics}

\item{xaxis_label}{the label for the x axis (e.g. Time (min))}

\item{yaxis_label}{the label for the y axis (e.g. Level (a.u.))}

\item{column_to_read}{the name of the column to process}
}
\description{
Plot model readouts with statistics for each time point.
}
