% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_sim.r
\name{get_column_names_statistics}
\alias{get_column_names_statistics}
\title{Return the column names of the statitics to calculate.}
\usage{
get_column_names_statistics(column.names, readout)
}
\arguments{
\item{column.names}{an array of column names}

\item{readout}{the name of the readout}
}
\value{
the column names including the readout name
}
\description{
Return the column names of the statitics to calculate.
}
