% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{get_param_names}
\alias{get_param_names}
\title{Get parameter names}
\usage{
get_param_names(filename)
}
\arguments{
\item{filename}{the filename containing the best fits}
}
\value{
the parameter names
}
\description{
Get parameter names
}
