% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_ps1.r
\name{get_sorted_level_indexes}
\alias{get_sorted_level_indexes}
\title{Return the indexes of the files as sorted by levels.}
\usage{
get_sorted_level_indexes(files)
}
\arguments{
\item{files}{the scanned files.}
}
\value{
the index of the levels
}
\description{
Return the indexes of the files as sorted by levels.
}
