% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_sim.r
\name{get_stats}
\alias{get_stats}
\title{Add the statistics for a readout to the table of statistics. The first column is Time.}
\usage{
get_stats(statistics, readout, colidx = 2)
}
\arguments{
\item{statistics}{the table of statistics to fill up}

\item{readout}{the statistics for this readout.}

\item{colidx}{the position in the table to put the readout statistics}
}
\value{
The table of statistics including this readout.
}
\description{
Add the statistics for a readout to the table of statistics. The first column is Time.
}
