% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_plots.r
\name{histogramplot}
\alias{histogramplot}
\title{Plot a generic histogram}
\usage{
histogramplot(dfCol, g = ggplot())
}
\arguments{
\item{dfCol}{a data frame with exactly one column.}

\item{g}{the current ggplot to overlap}
}
\value{
the plot
}
\description{
Plot a generic histogram
}
\examples{
a <- as.data.frame(rnorm(100))
histogramplot(dfCol=a)
}
