% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{insulin_receptor_2}
\alias{insulin_receptor_2}
\title{A stochastic model simulation}
\format{A data frame with 2 variables:
\describe{
\item{Time}{The time variable}
\item{IR_beta_pY1146}{The insulin receptor beta phosphorylated at pY1146}
}}
\usage{
insulin_receptor_2
}
\description{
A stochastic model simulation
}
\keyword{datasets}
