% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{insulin_receptor_all_fits}
\alias{insulin_receptor_all_fits}
\title{A parameter estimation data set including all the evaluated fits.}
\format{A data frame with 4 variables:
\describe{
\item{ObjectiveValue}{the objective value for this parameter set}
\item{k1}{First estimated parameter: kinetic rate constant for IRbeta phosphorylation.}
\item{k2}{Second estimated parameter: kinetic rate constant for IRbeta refractory status.}
\item{k3}{Third estimated parameter: kinetic rate constant for IRbeta dephosphorylation.}
}}
\usage{
insulin_receptor_all_fits
}
\description{
Estimated parameters for a mini model of the insulin receptor (IRbeta).
}
\keyword{datasets}
