% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{insulin_receptor_IR_beta_pY1146}
\alias{insulin_receptor_IR_beta_pY1146}
\title{A stochastic simulation data set for the insulin receptor beta phosphorylated at pY1146.}
\format{A data frame with 41 variables:
\describe{
  \item{Time}{The time variable}
  \item{X1}{A stochastic simulation}
  \item{X2}{A stochastic simulation}
  \item{X3}{A stochastic simulation}
  \item{X4}{A stochastic simulation}
  \item{X5}{A stochastic simulation}
  \item{X6}{A stochastic simulation}
  \item{X7}{A stochastic simulation}
  \item{X8}{A stochastic simulation}
  \item{X9}{A stochastic simulation}
  \item{X10}{A stochastic simulation}
  \item{X11}{A stochastic simulation}
  \item{X12}{A stochastic simulation}
  \item{X13}{A stochastic simulation}
  \item{X14}{A stochastic simulation}
  \item{X15}{A stochastic simulation}
  \item{X16}{A stochastic simulation}
  \item{X17}{A stochastic simulation}
  \item{X18}{A stochastic simulation}
  \item{X19}{A stochastic simulation}
  \item{X20}{A stochastic simulation}
  \item{X21}{A stochastic simulation}
  \item{X22}{A stochastic simulation}
  \item{X23}{A stochastic simulation}
  \item{X24}{A stochastic simulation}
  \item{X25}{A stochastic simulation}
  \item{X26}{A stochastic simulation}
  \item{X27}{A stochastic simulation}
  \item{X28}{A stochastic simulation}
  \item{X29}{A stochastic simulation}
  \item{X30}{A stochastic simulation}
  \item{X31}{A stochastic simulation}
  \item{X32}{A stochastic simulation}
  \item{X33}{A stochastic simulation}
  \item{X34}{A stochastic simulation}
  \item{X35}{A stochastic simulation}
  \item{X36}{A stochastic simulation}
  \item{X37}{A stochastic simulation}
  \item{X38}{A stochastic simulation}
  \item{X39}{A stochastic simulation}
  \item{X40}{A stochastic simulation}
}}
\usage{
insulin_receptor_IR_beta_pY1146
}
\description{
Independent stochastic simulation time courses for the phosphorylated insulin receptor upon insulin stimulation. This data set is summarised.
}
\keyword{datasets}
