% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{insulin_receptor_ps1_l8}
\alias{insulin_receptor_ps1_l8}
\title{A deterministic simulation of the insulin receptor model upon scanning of
1 model parameter. The initial amount of IR-beta is approx 8.}
\format{A data frame with 3 variables:
\describe{
\item{Time}{The time variable.}
\item{IR_beta}{The unphosphorylated state of the insulin receptor beta. The scanned variable.}
\item{IR_beta_pY1146}{The insulin receptor beta phosphorylated at pY1146}
}}
\usage{
insulin_receptor_ps1_l8
}
\description{
A deterministic simulation of the insulin receptor model upon scanning of
1 model parameter. The initial amount of IR-beta is approx 8.
}
\keyword{datasets}
