% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_sim.r
\name{kurtosis}
\alias{kurtosis}
\title{Calculate the kurtosis of a numeric vector}
\usage{
kurtosis(x, na.rm = FALSE)
}
\arguments{
\item{x}{the numeric vector}

\item{na.rm}{TRUE if NA values should be discarded}
}
\value{
the kurtosis
}
\description{
Calculate the kurtosis of a numeric vector
}
\examples{
kurtosis(x=c(1,2.4,5,NA), na.rm=TRUE)
}
