% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{leftCI}
\alias{leftCI}
\title{Return the left value confidence interval.}
\usage{
leftCI(cut_dataset, full_dataset, objval_col_idx, param_col_idx,
  objval_conf_level)
}
\arguments{
\item{cut_dataset}{a subset of the full dataset (e.g. the best 66\%, the best 95\%)}

\item{full_dataset}{the full dataset}

\item{objval_col_idx}{the index for the objective function column in the dataset}

\item{param_col_idx}{the index for the parameter column in the dataset}

\item{objval_conf_level}{the objective function confidence level}
}
\value{
the left confidence interval
}
\description{
Return the left value confidence interval.
}
