% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_sim.r
\name{load_exp_dataset}
\alias{load_exp_dataset}
\title{Load the experimental data set.}
\usage{
load_exp_dataset(exp_dataset, plot_exp_dataset = FALSE)
}
\arguments{
\item{exp_dataset}{a full path file containing the experimental data.}

\item{plot_exp_dataset}{TRUE if the experimental data should also be plotted}
}
\value{
the loaded data set.
}
\description{
Load the experimental data set.
}
