% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_plots.r
\name{normalise_vec}
\alias{normalise_vec}
\title{Normalise a vector within 0 and 1}
\usage{
normalise_vec(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{the vector to normalise}

\item{na.rm}{TRUE if NA values should be discarded}
}
\value{
the normalised vector
}
\description{
Normalise a vector within 0 and 1
}
\examples{
v <- c(-4,2,10,25,9,NA)
normalise_vec(vec=v)
}
