% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\docType{data}
\name{objval.col}
\alias{objval.col}
\title{The name of the Objective Value column}
\format{An object of class \code{character} of length 1.}
\usage{
objval.col
}
\description{
The name of the Objective Value column
}
\keyword{datasets}
