% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_ggplot2_themes.r
\name{pca_theme}
\alias{pca_theme}
\title{A generic basic theme for pca. It extends ggplot2 theme_classic().}
\usage{
pca_theme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{the font size}

\item{base_family}{the font family}
}
\description{
A generic basic theme for pca. It extends ggplot2 theme_classic().
}
\examples{
library(ggplot2)
theme_set(pca_theme(36))
}
