% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{plot_2d_cl_corr}
\alias{plot_2d_cl_corr}
\title{Plot parameter correlations using the 66\%, 95\%, or 99\% confidence level data sets}
\usage{
plot_2d_cl_corr(df66, df95, df99, objval_col, dfCols, plots_dir, model,
  plot_2d_66cl_corr = FALSE, plot_2d_95cl_corr = FALSE,
  plot_2d_99cl_corr = FALSE, logspace = TRUE, scientific_notation = TRUE)
}
\arguments{
\item{df66}{the data frame filtered at 66\%}

\item{df95}{the data frame filtered at 95\%}

\item{df99}{the data frame filtered at 95\%}

\item{objval_col}{the objective value column name}

\item{dfCols}{the column names of the dataset}

\item{plots_dir}{the directory to save the generated plots}

\item{model}{the model name}

\item{plot_2d_66cl_corr}{true if the 2D parameter correlation plots for 66\% confidence intervals should be plotted. This is time consuming. (default: FALSE)}

\item{plot_2d_95cl_corr}{true if the 2D parameter correlation plots for 95\% confidence intervals should be plotted. This is time consuming. (default: FALSE)}

\item{plot_2d_99cl_corr}{true if the 2D parameter correlation plots for 99\% confidence intervals should be plotted. This is time consuming. (default: FALSE)}

\item{logspace}{true if parameters should be plotted in logspace. (default: TRUE)}

\item{scientific_notation}{true if the axis labels should be plotted in scientific notation (default: TRUE)}
}
\description{
Plot parameter correlations using the 66\%, 95\%, or 99\% confidence level data sets
}
