% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_ps2.r
\name{plot_double_param_scan_data}
\alias{plot_double_param_scan_data}
\title{Plot model double parameter scan time courses.}
\usage{
plot_double_param_scan_data(model, scanned_par1, scanned_par2, inputdir,
  outputdir, run)
}
\arguments{
\item{model}{the model name without extension}

\item{scanned_par1}{the 1st scanned parameter}

\item{scanned_par2}{the 2nd scanned parameter}

\item{inputdir}{the input directory}

\item{outputdir}{the output directory}

\item{run}{the simulation run}
}
\description{
Plot model double parameter scan time courses.
}
