% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{plot_objval_vs_iters}
\alias{plot_objval_vs_iters}
\title{Plot the Objective values vs Iterations}
\usage{
plot_objval_vs_iters(objval_array, plots_dir, model)
}
\arguments{
\item{objval_array}{the array of objective function values.}

\item{plots_dir}{the directory to save the generated plots}

\item{model}{the model name}
}
\description{
Plot the Objective values vs Iterations
}
