% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{plot_parameter_correlations}
\alias{plot_parameter_correlations}
\title{Plot parameter correlations.}
\usage{
plot_parameter_correlations(df, dfCols, plots_dir, plot_filename_prefix,
  title = "", objval_col_idx = 1, logspace = TRUE,
  scientific_notation = TRUE)
}
\arguments{
\item{df}{the data frame}

\item{dfCols}{the columns of the data frame. Each column is a parameter. Only parameters to the left of objval_col_idx are plotted.}

\item{plots_dir}{the directory for storing the plots}

\item{plot_filename_prefix}{the prefix for the plot filename}

\item{title}{the plot title (default: "")}

\item{objval_col_idx}{the index of the column containing the objective value (default: 1)}

\item{logspace}{true if the parameters should be plotted in logspace (default: TRUE)}

\item{scientific_notation}{true if the axis labels should be plotted in scientific notation (default: TRUE)}
}
\description{
Plot parameter correlations.
}
