% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{plot_sampled_ple}
\alias{plot_sampled_ple}
\title{Plot the sampled profile likelihood estimations (PLE)}
\usage{
plot_sampled_ple(df99, objval_col, cl66_objval, cl95_objval, cl99_objval,
  plots_dir, model, logspace = TRUE, scientific_notation = TRUE)
}
\arguments{
\item{df99}{the 99\% confidence level data frame}

\item{objval_col}{the objective value column name}

\item{cl66_objval}{the 66\% confidence level objective value}

\item{cl95_objval}{the 95\% confidence level objective value}

\item{cl99_objval}{the 99\% confidence level objective value}

\item{plots_dir}{the directory to save the generated plots}

\item{model}{the model name}

\item{logspace}{true if parameters should be plotted in logspace. (default: TRUE)}

\item{scientific_notation}{true if the axis labels should be plotted in scientific notation (default: TRUE)}
}
\description{
Plot the sampled profile likelihood estimations (PLE)
}
