% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_ps1.r
\name{plot_single_param_scan_data}
\alias{plot_single_param_scan_data}
\title{Plot model single parameter scan time courses}
\usage{
plot_single_param_scan_data(model, inhibition_only, outputdir, sim_data_folder,
  sim_plots_folder, run, percent_levels = TRUE, min_level = 0,
  max_level = 100, levels_number = 10, xaxis_label = "",
  yaxis_label = "")
}
\arguments{
\item{model}{The model name}

\item{inhibition_only}{true if the scanning only decreases the variable amount (inhibition only)}

\item{outputdir}{the output directory}

\item{sim_data_folder}{the name of the folder containing the simulated data}

\item{sim_plots_folder}{the name of the folder containing the simulated plots}

\item{run}{the simulation number}

\item{percent_levels}{true if scanning levels are in percent (default TRUE)}

\item{min_level}{the minimum level (default: 0)}

\item{max_level}{the maximum level (default: 100)}

\item{levels_number}{the number of levels (default: 10)}

\item{xaxis_label}{the label for the x axis (e.g. Time (min))}

\item{yaxis_label}{the label for the y axis (e.g. Level (a.u.))}
}
\description{
Plot model single parameter scan time courses
}
