% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_ps1.r
\name{plot_single_param_scan_data_homogen}
\alias{plot_single_param_scan_data_homogen}
\title{Plot model single parameter scan time courses using homogeneous lines.}
\usage{
plot_single_param_scan_data_homogen(model, outputdir, sim_data_folder,
  sim_plots_folder, run, xaxis_label = "", yaxis_label = "")
}
\arguments{
\item{model}{The model name}

\item{outputdir}{the output directory}

\item{sim_data_folder}{the name of the folder containing the simulated data}

\item{sim_plots_folder}{the name of the folder containing the simulated plots}

\item{run}{the simulation number}

\item{xaxis_label}{the label for the x axis (e.g. Time (min))}

\item{yaxis_label}{the label for the y axis (e.g. Level (a.u.))}
}
\description{
Plot model single parameter scan time courses using homogeneous lines.
}
