% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{replace_colnames}
\alias{replace_colnames}
\title{Rename data frame columns. `ObjectiveValue` is renamed as `ObjVal`. Substrings `Values.` and `..InitialValue` are
removed.}
\usage{
replace_colnames(df.cols)
}
\arguments{
\item{df.cols}{The columns of a data frame.}
}
\value{
the renamed columns
}
\description{
Rename data frame columns. `ObjectiveValue` is renamed as `ObjVal`. Substrings `Values.` and `..InitialValue` are
removed.
}
