% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{sbpipe_objval_vs_iters_analysis}
\alias{sbpipe_objval_vs_iters_analysis}
\title{Analysis of the Objective values vs Iterations.}
\usage{
sbpipe_objval_vs_iters_analysis(model_name, filename, plots_dir)
}
\arguments{
\item{model_name}{the model name without extension}

\item{filename}{the filename containing the fits sequence}

\item{plots_dir}{the directory to save the generated plots}
}
\description{
Analysis of the Objective values vs Iterations.
}
