% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{sbpipe_pe_all_fits}
\alias{sbpipe_pe_all_fits}
\title{Run model parameter estimation analysis and plot results for all fits.}
\usage{
sbpipe_pe_all_fits(model, allfits_filenamein, plots_dir, data_point_num,
  fileout_param_estim_details, fileout_param_estim_summary, plot_2d_66cl_corr,
  plot_2d_95cl_corr, plot_2d_99cl_corr, logspace, scientific_notation)
}
\arguments{
\item{model}{the model name without extension.}

\item{allfits_filenamein}{the dataset containing all the parameter fits}

\item{plots_dir}{the directory to save the generated plots.}

\item{data_point_num}{the number of data points used for parameterise the model.}

\item{fileout_param_estim_details}{the name of the file containing the detailed statistics for the estimated parameters.}

\item{fileout_param_estim_summary}{the name of the file containing the summary for the parameter estimation.}

\item{plot_2d_66cl_corr}{true if the 2D parameter correlation plots for 66\% confidence intervals should be plotted.}

\item{plot_2d_95cl_corr}{true if the 2D parameter correlation plots for 95\% confidence intervals should be plotted.}

\item{plot_2d_99cl_corr}{true if the 2D parameter correlation plots for 99\% confidence intervals should be plotted.}

\item{logspace}{true if parameters should be plotted in logspace.}

\item{scientific_notation}{true if axis labels should be plotted in scientific notation.}
}
\description{
Run model parameter estimation analysis and plot results for all fits.
}
