% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{sbpipe_pe_ds_preproc}
\alias{sbpipe_pe_ds_preproc}
\title{Parameter estimation pre-processing. It renames the data set columns, and applies a log10 transformation if logspace is TRUE.
If all.fits is true, it also computes the confidence levels.}
\usage{
sbpipe_pe_ds_preproc(filename, param.names = c(), logspace = TRUE,
  all.fits = FALSE, data_point_num = 0,
  fileout_param_estim_summary = "param_estim_summary.csv")
}
\arguments{
\item{filename}{the dataset filename containing the fits sequence}

\item{param.names}{the list of estimated parameter names}

\item{logspace}{true if the data set shoud be log10-transformed.}

\item{all.fits}{true if filename contains all fits, false otherwise}

\item{data_point_num}{the number of data points used for parameterise the model. Ignored if all.fits is false}

\item{fileout_param_estim_summary}{the name of the file containing the summary for the parameter estimation. Ignored if all.fits is false}
}
\description{
Parameter estimation pre-processing. It renames the data set columns, and applies a log10 transformation if logspace is TRUE.
If all.fits is true, it also computes the confidence levels.
}
