% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{sbpipe_pe_final_fits}
\alias{sbpipe_pe_final_fits}
\title{Run model parameter estimation analysis and plot results for the best fits.}
\usage{
sbpipe_pe_final_fits(model, finalfits_filenamein, plots_dir, best_fits_percent,
  logspace, scientific_notation)
}
\arguments{
\item{model}{the model name without extension.}

\item{finalfits_filenamein}{the dataset containing the best parameter fits}

\item{plots_dir}{the directory to save the generated plots.}

\item{best_fits_percent}{the percent of best fits to analyse.}

\item{logspace}{true if parameters should be plotted in logspace.}

\item{scientific_notation}{true if axis labels should be plotted in scientific notation.}
}
\description{
Run model parameter estimation analysis and plot results for the best fits.
}
