% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_ps2.r
\name{sbpipe_ps2}
\alias{sbpipe_ps2}
\title{Main R function for SBpipe pipeline: parameter_scan2().}
\usage{
sbpipe_ps2(model_noext, scanned_par1, scanned_par2, inputdir, outputdir, run)
}
\arguments{
\item{model_noext}{the model name without extension}

\item{scanned_par1}{the 1st scanned parameter}

\item{scanned_par2}{the 2nd scanned parameter}

\item{inputdir}{the input directory}

\item{outputdir}{the output directory}

\item{run}{the simulation run}
}
\description{
Main R function for SBpipe pipeline: parameter_scan2().
}
