% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_sim.r
\name{sbpipe_sim}
\alias{sbpipe_sim}
\title{Main R function for SBpipe pipeline: simulate().}
\usage{
sbpipe_sim(model_noext, inputdir, outputdir, outputfile_stats,
  outputfile_repeats, exp_dataset, plot_exp_dataset, exp_dataset_alpha,
  xaxis_label, yaxis_label, column_to_read)
}
\arguments{
\item{model_noext}{the model name without extension}

\item{inputdir}{the input directory}

\item{outputdir}{the output directory}

\item{outputfile_stats}{the output file containing the statistics}

\item{outputfile_repeats}{the output file storing the model simulation repeats}

\item{exp_dataset}{the file containing the experimental data.}

\item{plot_exp_dataset}{TRUE if the experimental data should also be plotted}

\item{exp_dataset_alpha}{the alpha level for the data set}

\item{xaxis_label}{the label for the x axis (e.g. Time (min))}

\item{yaxis_label}{the label for the y axis (e.g. Level (a.u.))}

\item{column_to_read}{the name of the column to process}
}
\description{
Main R function for SBpipe pipeline: simulate().
}
