% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper.r
\docType{package}
\name{sbpiper-package}
\alias{sbpiper}
\alias{sbpiper-package}
\title{sbpiper: Data Analysis Functions for SBpipe Package}
\description{
Data analysis functions for mathematical
model simulation, parameter estimation, and parameter scan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pdp10/sbpiper}
  \item Report bugs at \url{https://github.com/pdp10/sbpiper/issues}
}

}
\author{
\strong{Maintainer}: Piero Dalle Pezze \email{piero.dallepezze@gmail.com}

Other contributors:
\itemize{
  \item Nicolas Le Novère \email{nicolas.lenovere@babraham.ac.uk} [copyright holder]
}

}
\keyword{internal}
