% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{sbpiper_objval_vs_iters_analysis}
\alias{sbpiper_objval_vs_iters_analysis}
\title{Analysis of the Objective values vs Iterations.}
\usage{
sbpiper_objval_vs_iters_analysis(model, filename, plots_dir)
}
\arguments{
\item{model}{the model name}

\item{filename}{the filename containing the fits sequence}

\item{plots_dir}{the directory to save the generated plots}
}
\description{
Analysis of the Objective values vs Iterations.
}
