% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_ps1.r
\name{sbpiper_ps1}
\alias{sbpiper_ps1}
\title{Main R function for SBpipe pipeline: parameter_scan1().}
\usage{
sbpiper_ps1(model, inhibition_only, inputdir, outputdir, run, percent_levels,
  min_level, max_level, levels_number, homogeneous_lines, xaxis_label,
  yaxis_label)
}
\arguments{
\item{model}{the model name}

\item{inhibition_only}{true if the scanning only decreases the variable amount (inhibition only)}

\item{inputdir}{the input directory containing the simulated data}

\item{outputdir}{the output directory that will contain the simulated plots}

\item{run}{the simulation number}

\item{percent_levels}{true if scanning levels are in percent}

\item{min_level}{the minimum level}

\item{max_level}{the maximum level}

\item{levels_number}{the number of levels}

\item{homogeneous_lines}{true if lines should be plotted homogeneously}

\item{xaxis_label}{the label for the x axis (e.g. Time (min))}

\item{yaxis_label}{the label for the y axis (e.g. Level (a.u.))}
}
\description{
Main R function for SBpipe pipeline: parameter_scan1().
}
