% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_sim.r
\name{skewness}
\alias{skewness}
\title{Calculate the skewness of a numeric vector}
\usage{
skewness(x, na.rm = FALSE)
}
\arguments{
\item{x}{the numeric vector}

\item{na.rm}{TRUE if NA values should be discarded}
}
\value{
the skewness
}
\description{
Calculate the skewness of a numeric vector
}
\examples{
skewness(x=c(1,2.4,5,NA), na.rm=TRUE)
}
