% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_sim.r
\name{summarise_data}
\alias{summarise_data}
\title{Summarise the simulations.}
\usage{
summarise_data(inputdir, model, outputfile, column_to_read = "X1")
}
\arguments{
\item{inputdir}{the input directory containing the time course files}

\item{model}{the model name}

\item{outputfile}{the name of the file to store the simulations}

\item{column_to_read}{the name of the column to process}
}
\description{
Summarise the simulations.
}
