% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_ggplot2_themes.r
\name{tc_theme}
\alias{tc_theme}
\title{A theme for time courses. It extends ggplot2 theme_classic().}
\usage{
tc_theme(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{the font size}

\item{base_family}{the font family}
}
\description{
A theme for time courses. It extends ggplot2 theme_classic().
}
\examples{
library(ggplot2)
theme_set(tc_theme(36))
}
