% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{compute_aic}
\alias{compute_aic}
\title{Compute the Akaike Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2}
\usage{
compute_aic(chi2 = 0, params = 0)
}
\arguments{
\item{chi2}{the Chi^2 for the model}

\item{params}{the number of model parameters}
}
\value{
the AIC
}
\description{
Compute the Akaike Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2
}
\examples{
compute_aic(chi2=10, params=5)
}
