% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{compute_aicc}
\alias{compute_aicc}
\title{Compute the corrected Akaike Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2}
\usage{
compute_aicc(chi2 = 0, params = 0, data_points = 0)
}
\arguments{
\item{chi2}{the Chi^2 for the model}

\item{params}{the number of model parameters}

\item{data_points}{the number of data points}
}
\value{
the AICc
}
\description{
Compute the corrected Akaike Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2
}
\examples{
compute_aicc(chi2=10, params=5, data_points=100)
}
