% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{compute_bic}
\alias{compute_bic}
\title{Compute the Bayesian Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2}
\usage{
compute_bic(chi2 = 0, params = 1, data_points = 1)
}
\arguments{
\item{chi2}{the Chi^2 for the model}

\item{params}{the number of model parameters}

\item{data_points}{the number of data points}
}
\value{
the BIC
}
\description{
Compute the Bayesian Information Criterion. Assuming additive Gaussian
measurement noise of width 1, the term -2ln(L(theta|y)) ~ SSR ~ Chi^2
}
\examples{
compute_bic(chi2=10, params=5, data_points=100)
}
