% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{compute_cl_objval}
\alias{compute_cl_objval}
\title{Compute the confidence level based on the minimum objective value.}
\usage{
compute_cl_objval(min_objval = 0, params = 1, data_points = 2,
  level = 0.05)
}
\arguments{
\item{min_objval}{the minimum objective value}

\item{params}{the number of parameters}

\item{data_points}{the number of data points}

\item{level}{the confidence level threshold (e.g. 0.01, 0.05)}
}
\value{
the confidence level based on minimum objective value
}
\description{
Compute the confidence level based on the minimum objective value.
}
\examples{
compute_cl_objval(min_objval=10, params=5, data_points=100)
}
