% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{compute_fratio_threshold}
\alias{compute_fratio_threshold}
\title{Compute the fratio threshold for the confidence level.}
\usage{
compute_fratio_threshold(params = 1, data_points = 2, level = 0.05)
}
\arguments{
\item{params}{the number of parameters}

\item{data_points}{the number of data points}

\item{level}{the confidence level threshold (e.g. 0.01, 0.05)}
}
\value{
the f-ratio threshold
}
\description{
Compute the fratio threshold for the confidence level.
}
\examples{
compute_fratio_threshold(params=5, data_points=100)
compute_fratio_threshold(params=5, data_points=100, level=0.01)
}
