% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{leftCI}
\alias{leftCI}
\title{Return the left value of the parameter confidence interval. The provided dataset has two columns: ObjVal | ParamValue}
\usage{
leftCI(smallest.param.value, full_dataset, objval_conf_level)
}
\arguments{
\item{smallest.param.value}{the smallest parameter value within the specified confidence level}

\item{full_dataset}{the full dataset}

\item{objval_conf_level}{the objective function confidence level}
}
\value{
the left confidence interval
}
\description{
Return the left value of the parameter confidence interval. The provided dataset has two columns: ObjVal | ParamValue
}
