% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_plots.r
\name{plot_fits}
\alias{plot_fits}
\title{Plot the number of iterations vs objective values in log10 scale.}
\usage{
plot_fits(objval.vec, g = ggplot())
}
\arguments{
\item{objval.vec}{the array of objective function values.}

\item{g}{the current ggplot to overlap}
}
\value{
the plot
}
\description{
Plot the number of iterations vs objective values in log10 scale.
}
\examples{
v <- 10*(rnorm(10000))^4 + 10
plot_fits(objval.vec=v) + basic_theme(36)
}
