% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{plot_objval_vs_iters}
\alias{plot_objval_vs_iters}
\title{Plot the Objective values vs Iterations}
\usage{
plot_objval_vs_iters(objval.vec, model_name, plots_dir)
}
\arguments{
\item{objval.vec}{the vector containing the objective values}

\item{model_name}{the model name without extension}

\item{plots_dir}{the directory to save the generated plots}
}
\description{
Plot the Objective values vs Iterations
}
