% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{plot_objval_vs_iters}
\alias{plot_objval_vs_iters}
\title{Plot the Objective values vs Iterations}
\usage{
plot_objval_vs_iters(objval.vec, model, plots_dir)
}
\arguments{
\item{objval.vec}{the vector containing the objective values}

\item{model}{the model name}

\item{plots_dir}{the directory to save the generated plots}
}
\description{
Plot the Objective values vs Iterations
}
\examples{
dir.create(file.path("pe_plots"))
v <- 10*(rnorm(10000))^4 + 10
plot_objval_vs_iters(objval.vec=v, model="model", plots_dir="pe_plots")
}
