% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_plots.r
\name{plot_repeated_tc}
\alias{plot_repeated_tc}
\title{Plot repeated time courses in the same plot separately. First column is Time.}
\usage{
plot_repeated_tc(df, g = ggplot(), title = "", xaxis_label = "",
  yaxis_label = "", alpha = 1, yaxis.min = NULL, yaxis.max = NULL)
}
\arguments{
\item{df}{a data frame}

\item{g}{the current ggplot to overlap}

\item{title}{the title of the plot}

\item{xaxis_label}{the xaxis label of the plot}

\item{yaxis_label}{the yaxis label of the plot}

\item{alpha}{the amount of alpha transparency}

\item{yaxis.min}{the lower limit for the y axis}

\item{yaxis.max}{the upper limit for the y axis}
}
\value{
the plot
}
\description{
Plot repeated time courses in the same plot separately. First column is Time.
}
