% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_ps1.r
\name{plot_single_param_scan_data_homogen}
\alias{plot_single_param_scan_data_homogen}
\title{Plot model single parameter scan time courses using homogeneous lines.}
\usage{
plot_single_param_scan_data_homogen(model, inputdir, outputdir, run,
  xaxis_label = "", yaxis_label = "")
}
\arguments{
\item{model}{The model name}

\item{inputdir}{the input directory containing the simulated data}

\item{outputdir}{the output directory that will contain the simulated plots}

\item{run}{the simulation number}

\item{xaxis_label}{the label for the x axis (e.g. Time (min))}

\item{yaxis_label}{the label for the y axis (e.g. Level (a.u.))}
}
\description{
Plot model single parameter scan time courses using homogeneous lines.
}
