% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_pe.r
\name{replace_colnames}
\alias{replace_colnames}
\title{Rename data frame columns. \code{ObjectiveValue} is renamed as \code{ObjVal}. Substrings \code{Values.} and \code{..InitialValue} are
removed.}
\usage{
replace_colnames(df.cols)
}
\arguments{
\item{df.cols}{The columns of a data frame.}
}
\value{
the renamed columns
}
\description{
Rename data frame columns. \code{ObjectiveValue} is renamed as \code{ObjVal}. Substrings \code{Values.} and \code{..InitialValue} are
removed.
}
