% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpipe_pe.r
\name{rightCI}
\alias{rightCI}
\title{Return the right value of the parameter confidence interval. The provided dataset has two columns: ObjVal | ParamValue}
\usage{
rightCI(largest.param.value, full_dataset, objval_conf_level)
}
\arguments{
\item{largest.param.value}{the largest parameter value within the specified confidence level}

\item{full_dataset}{the full dataset}

\item{objval_conf_level}{the objective function confidence level}
}
\value{
the right confidence interval
}
\description{
Return the right value of the parameter confidence interval. The provided dataset has two columns: ObjVal | ParamValue
}
