\name{sbpiper-package}
\alias{sbpiper-package}
\alias{sbpiper}
\docType{package}
\title{
\packageTitle{sbpiper}
}
\description{
\packageDescription{sbpiper}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sbpiper}
\packageIndices{sbpiper}
 An overview of how to use the package, including the most important 
 functions 
}
\author{
\packageAuthor{sbpiper}

Maintainer: \packageMaintainer{sbpiper}
}
\references{
 Literature or other references for background information 
}
 Optionally other standard keywords, one per line, from file KEYWORDS in 
 the R documentation directory 
\keyword{ package }
\seealso{
 Optional links to other man pages, e.g. 
 \code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
simple examples of the most important functions
}
