% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_plots.r
\name{scatterplot}
\alias{scatterplot}
\title{Plot a generic scatter plot}
\usage{
scatterplot(df, g = ggplot(), colNameX = "x", colNameY = "y",
  dot_size = 0.5)
}
\arguments{
\item{df}{a data frame}

\item{g}{the current ggplot to overlap}

\item{colNameX}{the name of the column for the X axis}

\item{colNameY}{the name of the column for the Y axis}

\item{dot_size}{the size of the dots in the scatterplot}
}
\value{
the plot
}
\description{
Plot a generic scatter plot
}
\examples{
df <- data.frame(a=rnorm(10000), b=rnorm(10000))
scatterplot(df, colNameX="a", colNameY="b")
}
