% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbpiper_plots.r
\name{scatterplot_log10}
\alias{scatterplot_log10}
\title{Plot a generic scatter plot in log10 scale}
\usage{
scatterplot_log10(df, g = ggplot(), colNameX = "x", colNameY = "y",
  dot_size = 0.5)
}
\arguments{
\item{df}{a data frame to trasform to log10 scale}

\item{g}{the current ggplot to overlap}

\item{colNameX}{the name of the column for the X axis}

\item{colNameY}{the name of the column for the Y axis}

\item{dot_size}{the size of the dots in the scatterplot}
}
\value{
the plot
}
\description{
Plot a generic scatter plot in log10 scale
}
\examples{
df <- data.frame(a=exp(rnorm(10000)), b=exp(rnorm(10000)))
scatterplot_log10(df, colNameX="a", colNameY="b")
}
